requirements.txt
# Discord Bot Requirements

# Core Discord library
discord.py>=2.3.0

# Database
aiosqlite>=0.19.0

# Environment variables
python-dotenv>=1.0.0

# Translation
deep-translator>=1.11.4

# System monitoring (for health check)
psutil>=5.9.0

# Genshin Impact integration
genshin>=1.7.0
cryptography>=41.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
